Import Inter font for use in charts

font_inter == "Inter"

font_inter_bold == "Inter-Bold"

font_inter_light == "Inter-Light"

import_inter()

font_inter

font_inter_bold

font_inter_light

Format

length 1 character vector

An object of class character of length 1.

An object of class character of length 1.

Note

This will take care of ensuring PDF/PostScript usage. The location of the font directory is displayed after the base import is complete. It is highly recommended that you install them on your system the same way you would any other font you wish to use in other programs.

font_inter_bold (a.k.a. "Inter-Bold") is not available on Windows and will throw a warning if used in plots.

font_inter_light (a.k.a. "Inter-Light") is not available on Windows and will throw a warning if used in plots.