iOS Gray Light Color Scales

These scales are based on light gray palette of iOS system colors. They have only been implemented in sfthemes because I wanted to bring all the color palettes mentioned in the Apple's HIG.

Warning:

iOS gray colors are not necessary suitable for general plotting; so consider your use-case carefully before adapting them. You can loosely compare them to ggplot2::scale_alpha().

scale_colour_ios_gray_light(accessible = FALSE, ...)

scale_fill_ios_gray_light(accessible = FALSE, ...)

Arguments

accessible
Indicates whether the accessible set of colors should be used

...
Arguments passed on to ggplot2::discrete_scale

aesthetics

The names of the aesthetics that this scale works with.

scale_name

The name of the scale that should be used for error messages associated with this scale.

palette

A palette function that when called with a single integer argument (the number of levels in the scale) returns the values that they should take (e.g., scales::hue_pal()).

name

The name of the scale. Used as the axis or legend title. If waiver(), the default, the name of the scale is taken from the first mapping used for that aesthetic. If NULL, the legend title will be omitted.

breaks

One of:

NULL for no breaks

waiver() for the default breaks (the scale limits)

A character vector of breaks

A function that takes the limits as input and returns breaks as output

labels

One of:

NULL for no labels

waiver() for the default labels computed by the transformation object

A character vector giving labels (must be same length as breaks)

A function that takes the breaks as input and returns labels as output

limits

One of:

NULL to use the default scale values

A character vector that defines possible values of the scale and their order

A function that accepts the existing (automatic) values and returns new ones

na.translate

Unlike continuous scales, discrete scales can easily show missing values, and do so by default. If you want to remove missing values from a discrete scale, specify na.translate = FALSE.

na.value

If na.translate = TRUE, what aesthetic value should the missing values be displayed as? Does not apply to position scales where NA is always placed at the far right.

drop

Should unused factor levels be omitted from the scale? The default, TRUE, uses the levels that appear in the data; FALSE uses all the levels in the factor.

guide

A function used to create a guide or its name. See guides() for more information.

super

The super class to use for the constructed scale

See also

Examples

if (FALSE) {

library(ggplot2)

ggplot(mtcars, aes(y = mpg, x = as.factor(cyl))) +
  geom_boxplot(aes(color = as.factor(cyl), fill = as.factor(cyl))) +
  labs(x = "Miles Per Gallon", y = "MPG", 
       fill = "Number of Cylinders",
       color = "Number of Cylinders",
       title = "Fuel Economy",
       subtitle = "Fuel economy declines as engines gets bigger",
       caption = "Data from the 1974 Motor Trend US magazine.") +
  theme_sf_light() +
  scale_fill_ios_gray_light() +
  scale_colour_ios_gray_light()
}